// services/core/types.ts
type ConversationStep =
  | 'welcome' // Step para escoger categoría
  | 'category_selection' // Step genérico - escoger items de categoría -> first_level_selection
  | 'item_selection' // Step genérico - escoger items de categoría -> second_level_selection
  | 'quantity_selection'
  | 'cart_actions'
  | 'checkout'
  | 'final';

export interface ConversationConfig {
  timeoutMinutes: number;
}

export interface BaseConversation {
  phoneNumber: string;
  step: ConversationStep;
  lastInteraction: number;
}

export type StepProps<T extends BaseConversation> = {
  phoneNumber: string;
  message: string;
  conversation: T;
};

export type StepHandler<T extends BaseConversation> = (Props: StepProps<T>) => Promise<string>;

export interface Conversation<T extends BaseConversation> {
  config: ConversationConfig;
  stepHandlers: Record<string, StepHandler<T>>;
}

export type InitialConvo<T> = Omit<T, 'phoneNumber' | 'lastInteraction'>;
// ...
export interface CartItem {
  name: string;
  quantity: number;
  price: number;
  category: string;
  itemIndex: number;
}

// Extender BaseConversation para incluir los campos específicos del carrito
export interface CartConversation extends BaseConversation {
  cart: CartItem[];
  selectedCategory?: string;
  selectedItem?: string;
  selectedItemIndex?: number;
}
// ...
export interface TenantInfo {
  name: string;
  transfersPhoneNumber: string;
  deliveryCost: number;
}
// ...
export type TenantHandler = StepHandler<CartConversation>;

export type GetWelcomeMessageFn = (msgPreliminar?: string, greeting?: boolean) => string;